First stab at allowing a playback to be aborted.
It only works if the first keystroke made is holding 2600 to shift to
playback mode and then the second being a key to play back a memory. At
this point, pressing and holding a key will abort playback after the
current one is finished. Subsequence playbacks are not aborted.  Nor are
playbacks aborted if one of the twelve other keys are pressed before shifting
into playback mode.